airflow/requirements.txt
cloudscraper
BeautifulSoup4
python-dotenv
pymysql
